use CGFloat instead of float
git-svn-id: svn+ssh://svn.savannah.nongnu.org/gap/trunk@3330 d25b4930-3292-e311-b074-0010c6de5da8

+++ libs/PDFKit/PDFDocument.m
@@ -168,7 +168,7 @@ static NSMutableDictionary *documentInfo = nil;
    {
       if (([self pageRotate: pageNum] == 90) || ([self pageRotate: pageNum] == 180))
       {
-         float height = size.height;
+         CGFloat height = size.height;
          size.height = size.width;
          size.width  = height;
       }
